requirements.txt
fastapi==0.135.1
langchain==1.2.13
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.20
langchain-openai==1.1.11
langchain-text-splitters==1.1.1
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pymilvus==2.6.10
python-dotenv==1.2.2
sentence-transformers==5.3.0
uvicorn==0.42.0
